move period-specific data access into separate file and make protocol configuration easy to re-use for more protocols
